Move Xpress and Halarose classes into a different file
