+++ polling_stations/apps/data_collection/management/commands/__init__.py
@@ -4,10 +4,12 @@ from data_collection.base_importers import (
     BaseCsvStationsJsonDistrictsImporter,
     BaseCsvStationsKmlDistrictsImporter,
     BaseCsvStationsCsvAddressesImporter,
-    BaseXpressCsvImporter,
-    BaseHalaroseCsvImporter,
     BaseShpStationsCsvAddressesImporter,
     BaseApiKmlStationsKmlDistrictsImporter,
     BaseApiShpZipStationsShpZipDistrictsImporter,
     BaseApiCsvStationsShpZipDistrictsImporter,
+)
+from data_collection.ems_importers import (
+    BaseXpressCsvImporter,
+    BaseHalaroseCsvImporter,
 )
